requirements.txt
openpyxl
xlrd
numpy
pandas
matplotlib
seaborn
requests
scikit-learn
transformers
accelerate
trl
auto-gptq
optimum
einops
bitsandbytes
sentence-transformers
prophet
huggingface_hub
python-dotenv
pypdf
python-docx
python-pptx
pillow
pytesseract
pathlib
langchain
#ollama-client
faiss-cpu
langchain-ollama
langchain_community
flask
plotly
langgraph